Добавлено удаление отзывов в ленту

+++ src/main/java/ru/yandex/practicum/filmorate/model/Feed.java
@@ -5,16 +5,13 @@ import lombok.Setter;
 import ru.yandex.practicum.filmorate.model.enums.EventTypes;
 import ru.yandex.practicum.filmorate.model.enums.OperationTypes;
 
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
-
 @Getter
 @Setter
 public class Feed {
 
     private Long eventId;
     private Long entityId;
-    private Long userId;
+    private int userId;
     private long timestamp;
     private EventTypes eventType;
     private OperationTypes operation;
